authorize for user document

+++ src/main/java/com/cmc/invitaservice/service/implement/DocumentServiceImplement.java
@@ -3,49 +3,67 @@ package com.cmc.invitaservice.service.implement;
 import com.cmc.invitaservice.models.external.request.CreateDocumentRequest;
 import com.cmc.invitaservice.models.external.request.UpdateDocumentRequest;
 import com.cmc.invitaservice.models.external.response.GetAllDocumentResponse;
+import com.cmc.invitaservice.repositories.ApplicationUserRepository;
 import com.cmc.invitaservice.repositories.InvitaDocumentRepository;
+import com.cmc.invitaservice.repositories.entities.ApplicationUser;
 import com.cmc.invitaservice.repositories.entities.InvitaDocument;
 import com.cmc.invitaservice.service.DocumentService;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Optional;
 
 @Service
 @Slf4j
 public class DocumentServiceImplement implements DocumentService {
     private InvitaDocumentRepository invitaDocumentRepository;
+    private ApplicationUserRepository applicationUserRepository;
 
     @Autowired
-    public DocumentServiceImplement(InvitaDocumentRepository invitaDocumentRepository){
+    public DocumentServiceImplement(InvitaDocumentRepository invitaDocumentRepository,
+                                    ApplicationUserRepository applicationUserRepository){
         this.invitaDocumentRepository = invitaDocumentRepository;
+        this.applicationUserRepository = applicationUserRepository;
     }
 
     @Override
     public GetAllDocumentResponse getAllDocument(){
-        List<InvitaDocument> invitaDocumentList=invitaDocumentRepository.findAll();
-
+        UserDetails userDetails = (UserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        String username = userDetails.getUsername();
+        List<InvitaDocument> invitaDocumentList = invitaDocumentRepository.findInvitaDocumentByApplicationUserUsername(username);
         GetAllDocumentResponse getAllDocumentResponse = new GetAllDocumentResponse();
         getAllDocumentResponse.setListDocument(invitaDocumentList);
-
         return getAllDocumentResponse;
     }
 
     @Override
     public void deleteDocument(Long id){
-        invitaDocumentRepository.deleteById(id);
+        InvitaDocument invitaDocument = invitaDocumentRepository.findInvitaDocumentById(id);
+        UserDetails userDetails = (UserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        String username = userDetails.getUsername();
+        if (invitaDocument.getApplicationUser().getUsername().equals(username)) invitaDocumentRepository.deleteById(id);
     }
 
     @Override
-    public Optional<InvitaDocument> getDocumentById(Long documentId){
-        return invitaDocumentRepository.findInvitaDocumentById(documentId);
+    public InvitaDocument getDocumentById(Long documentId){
+        InvitaDocument invitaDocument = invitaDocumentRepository.findInvitaDocumentById(documentId);
+        UserDetails userDetails = (UserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        String username = userDetails.getUsername();
+        if (invitaDocument.getApplicationUser().getUsername().equals(username))
+            return invitaDocumentRepository.findInvitaDocumentById(documentId);
+        return null;
     }
 
     @Override
     public InvitaDocument addDocument(CreateDocumentRequest createDocumentRequest){
+        UserDetails userDetails = (UserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        String username = userDetails.getUsername();
+        ApplicationUser applicationUser = applicationUserRepository.findByUsername(username);
         InvitaDocument invitaDocument = new InvitaDocument();
+        invitaDocument.setApplicationUser(applicationUser);
         invitaDocument.setCreateDocumentRequest(createDocumentRequest);
         invitaDocumentRepository.save(invitaDocument);
         return invitaDocument;
@@ -53,8 +71,12 @@ public class DocumentServiceImplement implements DocumentService {
 
     @Override
     public void changeDocument(UpdateDocumentRequest updateDocumentRequest, Long documentId){
-        InvitaDocument invitaDocument = invitaDocumentRepository.findInvitaDocumentById(documentId).get();
-        invitaDocument.setUpdateDocumentRequest(updateDocumentRequest);
-        invitaDocumentRepository.save(invitaDocument);
+        InvitaDocument invitaDocument = invitaDocumentRepository.findInvitaDocumentById(documentId);
+        UserDetails userDetails = (UserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        String username = userDetails.getUsername();
+        if (invitaDocument.getApplicationUser().getUsername().equals(username)) {
+            invitaDocument.setUpdateDocumentRequest(updateDocumentRequest);
+            invitaDocumentRepository.save(invitaDocument);
+        }
     }
 }
